moved AbstractCodeGenerationWorkflowRunConfiguration to org.palladiosimulator.analyzer.workflow.core
